split up files and changed translation comp usage
