+++ indico/modules/events/registration/client/js/components/participant-list/ParticipantList.module.scss
