+++ indico/modules/events/registration/client/js/components/participant-list/ParticipantList.tsx
@@ -1,10 +1,15 @@
+import React from 'react';
+
+import ParticipantAccordion from './ParticipantAccordion';
+import {ParticipantAccordionProps} from './types';
+
 export default function ParticipantListAccordion({
   published,
   totalParticipantCount,
   tables,
-}: ParticipantListAccordionProps) {
+}: ParticipantAccordionProps) {
   return (
-    <ParticipantListAccordion
+    <ParticipantAccordion
       published={published}
       totalParticipantCount={totalParticipantCount}
       tables={tables}
